xlwingspro_requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
cryptography==3.2.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
distlib==0.3.1
entrypoints==0.3
eventemitter==0.2.0
filelock==3.0.12
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-configuration==0.8.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
refinitiv-dataplatform==1.0.0a5
requests==2.24.0
requests-async==0.6.2
rfc3986==1.4.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wincertstore==0.2
wrapt==1.12.1
xlwings==0.20.7
zipp==3.3.0
zope.interface==5.1.2
